refactor(tvix/store): bump CalculateNARResponse.nar_size to u64
Reviewed-on: https://cl.tvl.fyi/c/depot/+/8591
Tested-by: BuildkiteCI

+++ store/src/nar/non_caching_calculation_service.rs
@@ -36,7 +36,7 @@ impl<BS: BlobService, DS: DirectoryService> NARCalculationService
         self.nar_renderer.write_nar(&mut cw, root_node)?;
 
         Ok(proto::CalculateNarResponse {
-            nar_size: cw.count() as u32,
+            nar_size: cw.count() as u64,
             nar_sha256: cw.into_inner().finalize().to_vec(),
         })
     }
